Rename component
The component should instead be named `PageTitle` to better indicate it's purpose as the only `h1` on a page and discourage multiple uses of it on a page
