+++ examples/Headings/PageTitle.md
@@ -1,4 +1,4 @@
 A top level heading, used for page headers, application titles etc.
 ```jsx
-<HeadingOne>Hello World</HeadingOne>
+<PageTitle>Hello World</PageTitle>
 ```
